Add Apache License 2.0
:Release Notes:
Added the Apache License 2.0 to the project repository.

:Detailed Notes:
This commit introduces the Apache License 2.0 file to the project.
This license outlines terms for use, reproduction, and distribution
of the project's work. It includes sections on definitions, grant of
copyright and patent licenses, redistribution, and liability. The file
is critical for legal clarity and compliance with open-source norms.

:Testing Performed:
Local Test: OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-15606] Fix Security Errors and Wno errors - pmklogd

Change-Id: I797b4371a822bedfb16268cbc9cb50abcd2ee7a8