Rename license file for clarity
:Release Notes:
Renamed the license file from Apache-2.0 to LICENSE.

:Detailed Notes:
The license file was renamed from Apache-2.0 to LICENSE to adhere to
common conventions and improve clarity. This change does not affect
the content of the license itself, only the filename. This renaming
helps in recognizing the file as a license document in the project.

:Testing Performed:
Local Test: OK

:QA Notes:
N/A

:Issues Addressed:
[WRR-15606] Fix Security Errors and Wno errors - pmklogd

Change-Id: I970d7efa46d233544c8adc40a3c0763c8393e3a2